Keep internals internal
