+++ internal/errors/errors.go
@@ -1,0 +1,9 @@
+package errors
+
+import (
+	"errors"
+)
+
+var (
+	ErrInternalNoArgumentsForTag = errors.New("no arguments for tag found")
+)
